requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
colorama==0.4.4
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-phonenumber-field==5.2.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
httpie==2.5.0
idna==3.2
numpy==1.21.2
oauthlib==3.1.1
phonenumbers==8.12.35
Pillow==8.3.2
pycparser==2.20
Pygments==2.10.0
PyJWT==2.3.0
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
